requirements.txt
langchain
openai
python-dotenv
requests
tiktoken